backend/requirements.txt
fastapi
uvicorn
python-multipart
manga-ocr
opencv-python
Pillow
numpy<2
deep-translator
torch
requests
ultralytics
huggingface_hub
anthropic
langdetect
segmentation-models-pytorch
albumentations
timm
pytesseract
pymupdf
